+++ src/views/Search/components/Result/ResultContainer.tsx
@@ -2,46 +2,34 @@ import { css } from '@emotion/react';
 import { memo, MutableRefObject, useRef, useState } from 'react';
 
 import { getBarrierFreeInfo, getSearchKeyword } from '@/apis/public/search';
-import { BigInfoIcon } from '@/assets/icon';
-import { DefaultImage } from '@/assets/image';
 import Loading from '@/components/Loading';
 import PageLoading from '@/components/PageLoading';
-import PlaceCard from '@/components/PlaceCard';
-import { MAP_FACILITIES_API_KEY } from '@/constants/facilities';
 import { useInfiniteScroll } from '@/hooks/use-infinite-scroll';
-import { COLORS, FONTS } from '@/styles/constants';
 import { SearchItem } from '@/types/search';
 
-import {
-  getFilterList,
-  INITIAL_FILTER_INDEX_INFO,
-} from '../../constants/category';
+import { INITIAL_FILTER_INDEX_INFO } from '../../constants/category';
 import { FilterFacilities, filterState } from '../../types/category';
+import RenderResult from './RenderResult';
 
-interface SearchResultProps {
+interface ResultContainerProps {
   searchWord: string;
   filterState: filterState;
   heartList: number[];
 }
 
-const SearchResult = memo((props: SearchResultProps) => {
-  const { searchWord, filterState, heartList } = props;
+const ResultContainer = memo((props: ResultContainerProps) => {
+  const { searchWord, ...restProps } = props;
 
   const [loading, setLoading] = useState(false);
 
   const [placeList, setPlaceList] = useState<Record<string, SearchItem>>({});
-
   const [filterIndexInfo, setFilterIndexInfo] = useState<
     Record<FilterFacilities, string[]>
   >(INITIAL_FILTER_INDEX_INFO);
 
-  // 무한스크롤
-  const options = {
-    root: null,
-    rootMargin: '0px',
-    threshold: 0,
-  };
+  const placeListRef = useRef<HTMLUListElement>(null);
 
+  // 무한스크롤
   const handleObserver = async (
     observer: IntersectionObserver,
     target: MutableRefObject<HTMLElement | null>,
@@ -56,11 +44,13 @@ const SearchResult = memo((props: SearchResultProps) => {
         numOfRows: 50,
         MobileOS: 'ETC',
         keyword: searchWord || '',
-        contentTypeId: 12,
+        contentTypeId: 14,
       });
 
       if (items === '') {
-        if (pageNo === 0) setPlaceList({});
+        if (pageNo === 0) {
+          setPlaceList({});
+        }
         target.current && observer.unobserve(target.current);
       } else {
         const newPlaceList = items.item.reduce(
@@ -96,8 +86,8 @@ const SearchResult = memo((props: SearchResultProps) => {
           }
         });
         setFilterIndexInfo(updatedFilterIndexInfo);
-
         setPlaceList((prev) => ({ ...prev, ...newPlaceList }));
+
         page.current++;
       }
     } finally {
@@ -106,27 +96,9 @@ const SearchResult = memo((props: SearchResultProps) => {
   };
 
   const targetElement = useInfiniteScroll({
-    options,
     handleObserver,
-    deps: [searchWord],
+    deps: [],
   });
-
-  const placeListRef = useRef<HTMLUListElement>(null);
-
-  const filterList = getFilterList(filterState);
-
-  const renderPlaceList =
-    filterList.length > 0
-      ? Array.from(
-          filterList.reduce((acc, filter, idx) => {
-            if (idx === 0) return acc;
-            const curSet = new Set(
-              filterIndexInfo[MAP_FACILITIES_API_KEY[filter]],
-            );
-            return acc.intersection(curSet);
-          }, new Set(filterIndexInfo[MAP_FACILITIES_API_KEY[filterList[0]]])),
-        )
-      : Object.keys(placeList);
 
   return (
     <>
@@ -134,27 +106,12 @@ const SearchResult = memo((props: SearchResultProps) => {
       {loading && Object.keys(placeList).length === 0 && <PageLoading />}
 
       <ul css={containerCss(Object.keys(placeList).length)} ref={placeListRef}>
-        {!loading && renderPlaceList.length === 0 ? (
-          <NoResultView />
-        ) : (
-          renderPlaceList.map((contentid, idx) => {
-            const { title, addr1, addr2, firstimage, firstimage2 } =
-              placeList[contentid];
-            return (
-              <li key={contentid}>
-                <PlaceCard
-                  idx={idx}
-                  contentid={contentid}
-                  placeName={title}
-                  address={addr1 + addr2}
-                  imgSrc={firstimage || firstimage2 || DefaultImage}
-                  isHeart={heartList.includes(Number(contentid))}
-                  buttonDisabled
-                />
-              </li>
-            );
-          })
-        )}
+        <RenderResult
+          loading={loading}
+          filterIndexInfo={filterIndexInfo}
+          placeList={placeList}
+          {...restProps}
+        />
         <div ref={targetElement} css={lastTargetCss(loading)} />
 
         {/* 무한스크롤 로딩 */}
@@ -164,21 +121,9 @@ const SearchResult = memo((props: SearchResultProps) => {
   );
 });
 
-const NoResultView = () => (
-  <div css={noResultContainerCss}>
-    <BigInfoIcon />
-    <div css={noResultTitleCss}>검색 결과가 없어요</div>
-    <p css={noResultInfoCss}>
-      검색 필터를 바꾸거나
-      <br />
-      다른 여행지를 검색해보세요!
-    </p>
-  </div>
-);
+ResultContainer.displayName = 'ResultContainer';
 
-SearchResult.displayName = 'SearchResult';
-
-export default SearchResult;
+export default ResultContainer;
 
 const containerCss = (placeLength: number) => css`
   display: flex;
@@ -197,27 +142,4 @@ const lastTargetCss = (loading: boolean) => css`
 
   width: 100%;
   height: 1px;
-`;
-
-const noResultContainerCss = css`
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-
-  margin: 6rem 0 1.2rem;
-`;
-
-const noResultTitleCss = css`
-  margin: 2rem 0 0.8rem;
-
-  color: ${COLORS.gray9};
-  text-align: center;
-
-  ${FONTS.Body2};
-`;
-
-const noResultInfoCss = css`
-  color: ${COLORS.brand1};
-  text-align: center;
-  ${FONTS.Small1};
 `;
